requirements.txt
# MCP package - requires Python >=3.10
mcp>=1.0.0

# Google ADK (Agent Development Kit) - includes vertexai
google-adk[vertexai]>=1.0.0

# FastAPI + ASGI server for Cloud Run
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# HTTP clients
httpx>=0.27.0
httpx-sse>=0.4.0
requests>=2.31.0